Fixed some files and reorganized some. There is a circular import issue though.
