+++ app/controllers/main_routes/laborStatusForm.py
@@ -20,7 +20,8 @@ from app.logic.userInsertFunctions import*
 from app.models.supervisor import Supervisor
 from app.logic.tracy import Tracy
 from app.controllers.main_routes.laborReleaseForm import createLaborReleaseForm
-from app.controllers.admin_routes.allPendingForms import saveStatus
+from app.logic.allPendingForms import saveStatus
+from app.logic.statusFormFunctions import createLaborStatusForm, checkForSecondLSFBreak
 
 @main_bp.route('/laborstatusform', methods=['GET'])
 @main_bp.route('/laborstatusform/<laborStatusKey>', methods=['GET'])
